requirements.txt
fastapi
uvicorn
torch
torchvision
torchaudio
transformers==4.49.0
sentence-transformers
datasets
nltk
spacy
Pillow
opencv-python-headless
matplotlib
scipy
inflect
cachetools
open-clip-torch
ultralytics
accelerate
python-multipart